Extract SteamEvent from greenworks_api.cc into a separate file.
